Make gene track plot

Make gene track plot

finemapping_annot_trackplot(
  finemapstats,
  region,
  gene.annots,
  bigSNP,
  txdb,
  genome = c("hg19", "hg38"),
  genetrack_db = c("txdb", "gene.annots", "UCSC"),
  filter_protein_coding_genes = TRUE,
  countsdata,
  peaks,
  HiC_loops,
  filter_HiCloops_genes = NULL,
  filter_HiCloops_snps = NULL,
  data_colors = seq_along(countsdata),
  data_ylim = c(0, 1),
  color_pip_by = c("locus", "cs", "none"),
  highlight_snps = NULL,
  highlight_colors = "pink",
  genelabel_side = c("above", "right", "left", "below"),
  track.sizes = NULL,
  rotation.title = 90,
  verbose = FALSE
)

Arguments

finemapstats

A GRanges object or data frame of finemapping summary statistics

region

A GRanges object or data frame for the genomic range to plot

gene.annots

A GRanges object of gene annotations

bigSNP

A bigsnpr object attached via bigsnpr::snp_attach()

txdb

A txdb object of gene annotations

genome

Genome assembly version, hg19 (default) or hg38.

genetrack_db

Select a gene annotation database to use. Options: txdb: use the txdb objec. gene.annots: use the gene.annots object. UCSC uses UCSC knownGene annotations.

filter_protein_coding_genes

If TRUE, only shows protein coding gene

countsdata

A list of counts data

peaks

A list of peaks

HiC_loops

A list of HiC loops, e.g. PC-HiC, ABC, etc.

filter_HiCloops_genes

If TRUE, only shows HiC loops connected to the gene(s)

filter_HiCloops_snps

If TRUE, only shows HiC loops connected to the SNP(s)

data_colors

Colors for the countsdata tracks

data_ylim

ylim range for the countsdata tracks

color_pip_by

color SNPs in the PIP track by locus, cs, or none (same color).

highlight_snps

SNPs (rsIDs) to highlight

highlight_colors

Colors for the highlighted SNPs

genelabel_side

Side to put gene labels, options are: above (default), right, left, below

track.sizes

Sizes of the tracks

rotation.title

Rotation of the track titles

verbose

if TRUE, print detail messages for plotting