Make gene manhattan plot

Make gene manhattan plot

gene_manhattan_plot(
  gene.pip.res,
  chr = "chr",
  pos = "pos",
  gene_name = "gene_name",
  gene_pip = "gene_pip",
  sig.pip = 0.8,
  highlight = TRUE,
  ylim = 1.25,
  point.size = 2,
  label.size = point.size * 2,
  font.size = 15,
  max.overlaps = 10,
  title = ""
)

Arguments

gene.pip.res

Gene level finemapping result

chr

Name of the chr column in the gene finemapping summary statistics data

pos

Name of the pos column in the gene finemapping summary statistics data

gene_name

Name of the gene name column in the gene finemapping summary statistics data

gene_pip

Name of the gene PIP column in the gene finemapping summary statistics data

sig.pip

Signficant gene PIP (default: 0.8)

highlight

Highlight genes with gene PIP > sig.pip

ylim

Truncate gene PIP to ylim value in the plot (default: 1.25).

point.size

Size of the points.

label.size

Size of the labels.

font.size

Font size of the text.

max.overlaps

Exclude text labels that overlap too many things (default: 10).

title

Title of the plot