Cleans GWAS summary statistics and adds metadata

Cleans GWAS summary statistics and adds metadata

process_gwas_sumstats(
  sumstats_file,
  chr = "chr",
  pos = "pos",
  beta = "beta",
  se = "se",
  a0 = "a0",
  a1 = "a1",
  snp = "snp",
  pval = "pval",
  bigSNP,
  LD_Blocks
)

Arguments

sumstats_file

file containing raw summary statistics. Coordinates should be hg19/b37; we current do not support hg38.

chr

Name of the chromosome column in summary statistics file.

pos

Name of the position column (base pair position, in hg19!).

beta

Name of beta column (if you have Odds Ratio, you will need to transform it to log(Odds Ratio)).

se

Name of the standard error (se) column.

a0

Column name of the reference allele.

a1

Column name of the association/effect allele.

snp

Name of the SNP ID (rsID) column.

pval

Name of the p-value column.

bigSNP

bigSNP object from bigsnpr.

LD_Blocks

Reference LD blocks

Value

Cleaned summary statistics + LD block of every SNP, as well as its index in the reference panel of genotypes