Make gene track plot using Gviz

Make gene track plot using Gviz

track_plot(
  finemapstats,
  region,
  gene.annots,
  bigSNP,
  txdb,
  counts,
  peaks,
  loops,
  genome = "hg19",
  filter_loop_genes = NULL,
  filter_loop_snps = NULL,
  filter_protein_coding_genes = TRUE,
  color_piptrack_by = c("locus", "cs", "none"),
  highlight_snps = NULL,
  counts.ylim = c(0, 1),
  counts.color,
  peaks.color = "navy",
  loops.color = "gray",
  highlight.color = "pink",
  genelabel.side = c("above", "below", "left", "right"),
  track.sizes,
  rotation.title = 0,
  verbose = FALSE,
  ...
)

Arguments

finemapstats

A GRanges object of processed finemapping summary statistics

region

The genomic region to visualize in the format of "chr:start-end".

gene.annots

A GRanges object of gene annotations

bigSNP

A bigsnpr object attached via bigsnpr::snp_attach()

txdb

A txdb object of gene annotations

counts

A list of counts data to display as quantitativ data tracks

peaks

A list of peaks to display as binary data tracks

loops

A list of chromatin loops, e.g. PC-HiC, ABC, etc.

genome

Genome assembly version (default: "hg19").

filter_loop_genes

A vector of gene names. Only show loops connected to the genes.

filter_loop_snps

A vector of SNP IDs. Only show loops connected to the SNPs.

filter_protein_coding_genes

Logical. If TRUE, only shows protein coding gene.

color_piptrack_by

Color SNPs in the PIP track by locus, cs (credible sets), or none (same color).

highlight_snps

SNPs (rsIDs) to highlight. Or highlight top SNP ('topSNP').

counts.ylim

ylim range (default: between 0 and 1) for the counts tracks

counts.color

Colors for the counts tracks

peaks.color

Colors for the peak tracks

highlight.color

Colors for the highlighted SNPs

genelabel.side

Side to put gene labels, options are: 'above' (default), 'below', 'left', 'right'.

track.sizes

Sizes of the tracks

rotation.title

The rotation angle for the text in the track title

verbose

If TRUE, print detail messages for plotting